third_party/WebKit/LayoutTests/platform/mac/virtual/webrtc-wpt-unified-plan/external/wpt/webrtc/RTCPeerConnection-setRemoteDescription-nomsid-expected.txt

This is a testharness.js-based test.
FAIL setRemoteDescription with an SDP without a=msid lines triggers ontrack with a default stream. assert_equals: expected 1 but got 0
Harness: the test ran to completion.
